feat: support package mode (#7)
