+++ pyproject.toml
@@ -3,7 +3,10 @@ name = "fortress"
 version = "0.1.0"
 description = "for fetching secrets from Hashicorp Vault"
 authors = ["Artsy Product Sapphire Team <sapphire@artsymail.com>"]
-package-mode = false
+
+packages = [
+    { include = "fortress", from = "src" },
+]
 
 [tool.poetry.dependencies]
 boto3 = "^1.26.61"
